Create `Cup` class and draw multiple cups
Create a `Cup` class with `draw` method that calls `drawConicalFrustum`
function. Also add parameters to `drawConicalFrustum` function to allow
for top and/or bottom to be left open (i.e. not drawn). Draw multiple
cups by creating a `Cup` object for each cup and calling the `draw`
method of each object.