Add `normalPosition` field to `Cup` class
Add `normalPosition` field to `Cup` class, which stores the cup's
position when an animation is not being run. Animations change the
`position` vector, which is the cup's position in the window relative to
this `normalPosition` vector. The cup's `draw` method translates by the
`position` and `normalPosition` vectors before actually drawing the cup
shape. Previously the same concept was in place, but by translating the
window by the same vector instead of storing it in a field.