Allow >1 property to be animated simultaneously
Change to allow more than one property to be animated at a time, by
adding an `AnimationGroup` class, so, for example, the x- and
z-coordinates of a cup's normal position can be animated simultaneously.
This will allow cups to perform the "swap" animation without appearing
to go through each other, although for this to work I still need to
change animations so that the components of the vector not being
animated (e.g. the x- and z-components while the y-component is being
animated) aren't reset to zero.