Change so "swap" animation works properly
Change so "swap" animation works properly, by changing the `Animation`
class so that only the vector components being animated are changed and
the components not being animated aren't reset to zero, and so that
animations don't rely on a `PVector` object, by providing functions for
getting and setting the variable to be animated into the `Animation`
constructor. These functions are then called within the `Animation`
class when necessary.