Add ball
Change so ball appears under one of the cups (with the chosen cup being
randomly selected), by creating a `Ball` class, of which an instance is
passed into the `Cup` constructor if that cup contains the ball. The
ball moves with the cups when being swapped, but not when revealing.
Also remove unused `number` field from `Cup` class.